71-Password-Recovery Made:

+++ app/src/main/java/app/cashadvisor/common/domain/model/ErrorEntity.kt
@@ -66,7 +66,7 @@ sealed class ErrorEntity(open val message: String) {
         data class WrongConfirmationCode(
             override val message: String,
             val remainingAttempts:Int,
-            val lockDuration:Int
+            val lockDuration:Long
         ) : ConfirmResetPasswordByEmailWithCode(message)
 
         data class FailedToConfirmPasswordReset(override val message: String) :
