+++ app/src/main/java/app/cashadvisor/common/utill/exceptions/ResetPasswordException.kt
@@ -24,7 +24,7 @@ sealed class ResetPasswordException(
         class UnauthorizedInvalidTokenOrMissingContentTypeHeader(
             override val message: String,
             val remainingAttempts:Int,
-            val lockDuration:Int,
+            val lockDuration:Long,
             val statusCode: Int
         ):ConfirmResetPasswordByEmailWithCode(message = message)
